Producer/requirements.txt
kafka_python
beautifulsoup4
numpy
pandas
pymongo
requests
